+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.18
 
 - Added support for wheel events
+- Added sample project generated with create-react-app
 
 ## 1.0.17